test: o teste do prazo do turno passa a exercitar o laço de leitura
Com `timeout=0` o prazo estourava na abertura do stream: o provedor nem
chegava a ser chamado, o `stall` do dublê ficava inerte e apagar o prazo
de dentro do laço deixaria a suíte verde — o defeito voltaria sem sinal.

Agora o prazo é curto mas suficiente para abrir, e duas asserções travam
o regime: o stream precisa ter sido aberto e o iterador precisa ter sido
fechado. Um `wait_for` de fora garante que a ausência do prazo vire
falha em cinco segundos, e não uma suíte pendurada. Sem a correção do
código, este teste falha com TimeoutError. Um teste irmão cobre a outra
metade: prazo já estourado não gasta chamada ao provedor.

`chat_timeout_seconds` e `condense_timeout_seconds` passam a float, que
é como `asyncio.wait_for` os recebe — inteiros obrigariam o teste a
esperar um segundo para provar o que prova em cinquenta milissegundos.

Junto, as três sugestões abertas da avaliação: a fixture que força o
índice HNSW deixa de configurar o papel do banco e passa a usar o
`server_settings` da conexão, que nasce e morre com aquele pool; o
comentário de `test_chat_api.py` que dizia que `app.chat` não fecha o
iterador do provedor ficou obsoleto quando ele passou a fechar; e o
`.env.example` explica os dois prazos do chat.

+++ backend/app/config.py
@@ -44,8 +44,11 @@ class Settings(BaseSettings):
     # similaridade que separa positivas de negativas e justifica este corte.
     similarity_threshold: float = 0.625
     history_window: int = 6
-    chat_timeout_seconds: int = 60
-    condense_timeout_seconds: int = 5
+    # Segundos, e em ponto flutuante porque é assim que `asyncio.wait_for` os
+    # recebe: declará-los inteiros obrigaria o teste do prazo a esperar um
+    # segundo inteiro para provar que ele existe.
+    chat_timeout_seconds: float = 60.0
+    condense_timeout_seconds: float = 5.0
 
     postgres_user: str = "talkdoc"
     postgres_password: str = "talkdoc"
